working with my blog component
